refactor(jito-bell): extract program handlers into handlers/ module
