+++ jito-bell/src/lib.rs
@@ -1,16 +1,8 @@
-use std::{path::PathBuf, str::FromStr};
+use std::path::PathBuf;
 
 use borsh::BorshDeserialize;
-use defillama_rs::{
-    models::{Chain, Token},
-    DefiLlamaClient,
-};
 use error::JitoBellError;
 use futures::{sink::SinkExt, stream::StreamExt};
-use ix_parser::{
-    stake_pool::SplStakePoolProgram, token_2022::SplToken2022Program, vault::JitoVaultProgram,
-};
-use jito_vault_client::accounts::Vault;
 use log::{debug, error};
 use metrics::EpochMetrics;
 use solana_metrics::datapoint_info;
@@ -30,12 +22,7 @@ use yellowstone_grpc_proto::{
 };
 
 use crate::{
-    config::JitoBellConfig,
-    event_parser::{jito_steward::JitoStewardEvent, EventParser},
-    ix_parser::{jito_steward::JitoStewardInstruction, InstructionParser},
-    notification_info::Destination,
-    program::{EventConfig, Instruction, ProgramName},
-    tx_parser::JitoTransactionParser,
+    config::JitoBellConfig, notification_info::Destination, tx_parser::JitoTransactionParser,
 };
 
 pub mod cli_args;
@@ -43,6 +30,7 @@ pub mod config;
 mod error;
 pub mod event_parser;
 pub mod events;
+mod handlers;
 pub mod ix_parser;
 mod metrics;
 pub mod multi_writer;
@@ -94,7 +82,7 @@ impl JitoBellHandler {
     /// Sort thresholds
     ///
     /// - Sort values from high to low
-    fn sort_thresholds(&self, thresholds: &mut [ThresholdConfig]) {
+    pub(crate) fn sort_thresholds(&self, thresholds: &mut [ThresholdConfig]) {
         thresholds.sort_by(|a, b| {
             b.value
                 .partial_cmp(&a.value)
@@ -105,7 +93,7 @@ impl JitoBellHandler {
     /// Get divisor
     ///
     /// - Fetch Mint account to get decimals value, if fails return default 9
-    async fn divisor(&self, vrt: &Pubkey) -> f64 {
+    pub(crate) async fn divisor(&self, vrt: &Pubkey) -> f64 {
         let decimals = match self.rpc_client.get_account(vrt).await {
             Ok(mint_acc) => match Mint::unpack(&mint_acc.data) {
                 Ok(acc) => acc.decimals,
@@ -120,7 +108,7 @@ impl JitoBellHandler {
     /// Get VRT Symbol
     ///
     /// - Fetch Metadata account to get symbol value, if fails return default "VRT"
-    async fn vrt_symbol(&self, vrt: &Pubkey) -> String {
+    pub(crate) async fn vrt_symbol(&self, vrt: &Pubkey) -> String {
         let meta_pubkey =
             jito_vault_sdk::inline_mpl_token_metadata::pda::find_metadata_account(vrt).0;
         let symbol = match self.rpc_client.get_account(&meta_pubkey).await {
@@ -202,696 +190,10 @@ impl JitoBellHandler {
         Ok(())
     }
 
-    /// Send notification
-    pub async fn send_notification(
-        &mut self,
-        parser: &JitoTransactionParser,
-    ) -> Result<(), JitoBellError> {
-        for program in &parser.instructions {
-            match program {
-                InstructionParser::SplToken2022(_) => {
-                    debug!("Token 2022");
-                }
-                InstructionParser::SplStakePool(spl_stake_program) => {
-                    debug!("SPL Stake Pool");
-
-                    let spl_program_str = spl_stake_program.to_string();
-
-                    let instruction_opt = self
-                        .config
-                        .programs
-                        .get(&ProgramName::SplStakePool)
-                        .and_then(|program_config| {
-                            program_config.instructions.get(&spl_program_str).cloned()
-                        });
-
-                    if let Some(instruction) = instruction_opt {
-                        self.handle_spl_stake_pool_program(parser, spl_stake_program, &instruction)
-                            .await?;
-                    }
-                }
-                InstructionParser::JitoVault(jito_vault_program) => {
-                    debug!("Jito Vault");
-
-                    let jito_vault_program_str = jito_vault_program.to_string();
-
-                    let instruction_opt =
-                        self.config.programs.get(&ProgramName::JitoVault).and_then(
-                            |program_config| {
-                                program_config
-                                    .instructions
-                                    .get(&jito_vault_program_str)
-                                    .cloned()
-                            },
-                        );
-
-                    if let Some(instruction) = instruction_opt {
-                        self.handle_jito_vault_program(parser, jito_vault_program, &instruction)
-                            .await?;
-                    }
-                }
-                InstructionParser::JitoSteward(jito_steward_instruction) => {
-                    debug!("Jito Steward");
-
-                    let jito_steward_program_str = jito_steward_instruction.to_string();
-
-                    let instruction_opt = self
-                        .config
-                        .programs
-                        .get(&ProgramName::JitoSteward)
-                        .and_then(|program_config| {
-                            program_config
-                                .instructions
-                                .get(&jito_steward_program_str)
-                                .cloned()
-                        });
-
-                    if let Some(instruction) = instruction_opt {
-                        self.handle_jito_steward_program(
-                            parser,
-                            jito_steward_instruction,
-                            &instruction,
-                        )
-                        .await?;
-                    }
-                }
-            }
-        }
-
-        for event in &parser.events {
-            match event {
-                EventParser::JitoSteward(jito_steward_event) => {
-                    let jito_steward_event_str = jito_steward_event.to_string();
-
-                    let event_opt = self
-                        .config
-                        .programs
-                        .get(&ProgramName::JitoSteward)
-                        .and_then(|program_config| {
-                            program_config.events.get(&jito_steward_event_str).cloned()
-                        });
-
-                    if let Some(event_config) = event_opt {
-                        let (description, amount, unit) = match jito_steward_event {
-                            JitoStewardEvent::StateTransition(state_transition) => {
-                                let desc = format!(
-                                    "Steward state transition occurred: {} → {}",
-                                    state_transition.previous_state, state_transition.new_state
-                                );
-                                (desc, None, None)
-                            }
-                            JitoStewardEvent::Rebalance(rebalance) => {
-                                let (change_type, amount_lamports) =
-                                    if rebalance.increase_lamports > 0 {
-                                        ("Stake Increase", rebalance.increase_lamports)
-                                    } else {
-                                        (
-                                            "Stake Decrease",
-                                            rebalance.decrease_components.total_unstake_lamports,
-                                        )
-                                    };
-
-                                let amount_sol = amount_lamports as f64 / 1_000_000_000.0;
-                                let type_emoji = if rebalance.increase_lamports > 0 {
-                                    "📈"
-                                } else {
-                                    "📉"
-                                };
-                                let validator_full = rebalance.vote_account.to_string();
-                                let validator_url = format!(
-                                    "https://www.jito.network/stakenet/steward/{validator_full}/"
-                                );
-
-                                let desc = format!(
-                                    "{} *{}* | {:.2} SOL\n\
-                                    \n\
-                                    Validator: <{}|{}>\n\
-                                    Epoch: {} | Type: {:?}",
-                                    type_emoji,
-                                    change_type,
-                                    amount_sol,
-                                    validator_url,
-                                    validator_full,
-                                    rebalance.epoch,
-                                    rebalance.rebalance_type_tag
-                                );
-
-                                (desc, Some(amount_sol), Some("SOL"))
-                            }
-                            JitoStewardEvent::DirectedRebalance(rebalance) => {
-                                let (change_type, amount_lamports) =
-                                    if rebalance.increase_lamports > 0 {
-                                        ("Stake Increase", rebalance.increase_lamports)
-                                    } else {
-                                        ("Stake Decrease", rebalance.decrease_lamports)
-                                    };
-
-                                let amount_sol = amount_lamports as f64 / 1_000_000_000.0;
-                                let type_emoji = if rebalance.increase_lamports > 0 {
-                                    "📈"
-                                } else {
-                                    "📉"
-                                };
-
-                                let validator_full = rebalance.vote_account.to_string();
-                                let validator_url = format!(
-                                    "https://www.jito.network/stakenet/steward/{validator_full}/"
-                                );
-
-                                let desc = format!(
-                                    "{} *{}* | {:.2} SOL\n\
-                                    \n\
-                                    Validator: <{}|{}>\n\
-                                    Epoch: {} | Type: {:?}",
-                                    type_emoji,
-                                    change_type,
-                                    amount_sol,
-                                    validator_url,
-                                    validator_full,
-                                    rebalance.epoch,
-                                    rebalance.rebalance_type_tag
-                                );
-
-                                (desc, Some(amount_sol), Some("SOL"))
-                            }
-                            _ => {
-                                debug!("Unhandled event type: {:?}", jito_steward_event);
-                                ("Unknown event".to_string(), None, None)
-                            }
-                        };
-                        match event_config {
-                            EventConfig::WithThresholds { thresholds } => {
-                                if let Some(amt) = amount {
-                                    let matching_threshold = thresholds
-                                        .iter()
-                                        .filter(|t| amt >= t.value)
-                                        .max_by(|a, b| {
-                                            a.value
-                                                .partial_cmp(&b.value)
-                                                .unwrap_or(std::cmp::Ordering::Equal)
-                                        });
-
-                                    if let Some(threshold) = matching_threshold {
-                                        let final_desc =
-                                            if threshold.notification.description.is_empty() {
-                                                description.clone()
-                                            } else {
-                                                format!(
-                                                    "{}\n\n{}",
-                                                    threshold.notification.description, description
-                                                )
-                                            };
-                                        self.dispatch_platform_notifications(
-                                            &threshold.notification.destinations,
-                                            &final_desc,
-                                            Some(amt),
-                                            unit,
-                                            &parser.transaction_signature,
-                                        )
-                                        .await?;
-                                    }
-                                }
-                            }
-                            EventConfig::Simple {
-                                destinations,
-                                description: config_desc,
-                            } => {
-                                // Use config description if provided, otherwise use generated description
-                                let final_desc = if config_desc.is_empty() {
-                                    description
-                                } else {
-                                    format!("{}\n\n{}", config_desc, description)
-                                };
-
-                                self.dispatch_platform_notifications(
-                                    &destinations,
-                                    &final_desc,
-                                    amount,
-                                    unit,
-                                    &parser.transaction_signature,
-                                )
-                                .await?;
-                            }
-                        }
-                    }
-                }
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Handle SPL Stake Pool Program
-    ///
-    /// - Notify only once for the first matching threshold.
-    async fn handle_spl_stake_pool_program(
-        &mut self,
-        parser: &JitoTransactionParser,
-        spl_stake_program: &SplStakePoolProgram,
-        instruction: &Instruction,
-    ) -> Result<(), JitoBellError> {
-        debug!("SPL Stake Program: {}", spl_stake_program);
-
-        match spl_stake_program {
-            SplStakePoolProgram::IncreaseValidatorStake { ix, amount } => {
-                let stake_pool_info = &ix.accounts[0];
-                let _staker_info = &ix.accounts[1];
-                let _withdraw_authority_info = &ix.accounts[2];
-                let _validator_list_info = &ix.accounts[3];
-                let _reserve_stake_account_info = &ix.accounts[4];
-                let _maybe_ephemeral_stake_account_info = &ix.accounts[5];
-                let _validator_stake_account_info = &ix.accounts[6];
-                let _validator_vote_account_info = &ix.accounts[7];
-                let _clock_info = &ix.accounts[8];
-                let _rent_info = &ix.accounts[9];
-                let _stake_history_info = &ix.accounts[10];
-                let _stake_config_info = &ix.accounts[11];
-                let _system_program_info = &ix.accounts[12];
-                let _stake_program_info = &ix.accounts[13];
-
-                if let Some(mut stake_pools) = instruction.stake_pools.clone() {
-                    if let Some(alert_config) =
-                        stake_pools.get_mut(&stake_pool_info.pubkey.to_string())
-                    {
-                        self.sort_thresholds(alert_config.thresholds.as_mut());
-                        for threshold in alert_config.thresholds.iter() {
-                            if *amount > threshold.value {
-                                self.dispatch_platform_notifications(
-                                    &threshold.notification.destinations,
-                                    &threshold.notification.description,
-                                    Some(*amount),
-                                    Some("SOL"),
-                                    &parser.transaction_signature,
-                                )
-                                .await?;
-                                break;
-                            }
-                        }
-                    }
-                }
-            }
-            SplStakePoolProgram::DepositStake { ix } => {
-                let _stake_pool_info = &ix.accounts[0];
-                let _validator_list_info = &ix.accounts[1];
-                let _stake_deposit_authority_info = &ix.accounts[2];
-                let withdraw_authority_info = &ix.accounts[3];
-                let _stake_info = &ix.accounts[4];
-                let _validator_stake_account_info = &ix.accounts[5];
-                let _reserve_stake_account_info = &ix.accounts[6];
-                let dest_user_pool_info = &ix.accounts[7];
-                let _manager_fee_info = &ix.accounts[8];
-                let _referrer_fee_info = &ix.accounts[9];
-                let pool_mint_info = &ix.accounts[10];
-
-                if let Some(mut lsts) = instruction.lsts.clone() {
-                    if let Some(alert_config) = lsts.get_mut(&pool_mint_info.pubkey.to_string()) {
-                        for program in &parser.instructions {
-                            if let InstructionParser::SplToken2022(program) = program {
-                                match program {
-                                    SplToken2022Program::MintTo { ix, amount } => {
-                                        let mint_info = &ix.accounts[0];
-                                        let destination_account_info = &ix.accounts[1];
-                                        let owner_info = &ix.accounts[2];
-
-                                        if mint_info.pubkey.eq(&pool_mint_info.pubkey)
-                                            && destination_account_info
-                                                .pubkey
-                                                .eq(&dest_user_pool_info.pubkey)
-                                            && owner_info.pubkey.eq(&withdraw_authority_info.pubkey)
-                                        {
-                                            self.sort_thresholds(alert_config.thresholds.as_mut());
-                                            for threshold in alert_config.thresholds.iter() {
-                                                if *amount as f64 > threshold.value {
-                                                    self.dispatch_platform_notifications(
-                                                        &threshold.notification.destinations,
-                                                        &threshold.notification.description,
-                                                        Some(*amount as f64),
-                                                        Some("SOL"),
-                                                        &parser.transaction_signature,
-                                                    )
-                                                    .await?;
-                                                    break;
-                                                }
-                                            }
-
-                                            break;
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-            SplStakePoolProgram::WithdrawStake {
-                ix,
-                minimum_lamports_out,
-            } => {
-                let _stake_pool_info = &ix.accounts[0];
-                let _validator_list_info = &ix.accounts[1];
-                let _withdraw_authority_info = &ix.accounts[2];
-                let _stake_split_from = &ix.accounts[3];
-                let _stake_split_to = &ix.accounts[4];
-                let _user_stake_authority_info = &ix.accounts[5];
-                let _user_transfer_authority_info = &ix.accounts[6];
-                let _burn_from_pool_info = &ix.accounts[7];
-                let _manager_fee_info = &ix.accounts[8];
-                let pool_mint_info = &ix.accounts[9];
-
-                if let Some(mut lsts) = instruction.lsts.clone() {
-                    if let Some(alert_config) = lsts.get_mut(&pool_mint_info.pubkey.to_string()) {
-                        self.sort_thresholds(alert_config.thresholds.as_mut());
-                        for threshold in alert_config.thresholds.iter() {
-                            if *minimum_lamports_out >= threshold.value {
-                                self.dispatch_platform_notifications(
-                                    &threshold.notification.destinations,
-                                    &threshold.notification.description,
-                                    Some(*minimum_lamports_out),
-                                    Some("SOL"),
-                                    &parser.transaction_signature,
-                                )
-                                .await?;
-                                break;
-                            }
-                        }
-                    }
-                }
-            }
-            SplStakePoolProgram::DepositSol { ix, amount } => {
-                let _stake_pool_info = &ix.accounts[0];
-                let _withdraw_authority_info = &ix.accounts[1];
-                let _reserve_stake_account_info = &ix.accounts[2];
-                let _from_user_lamports_info = &ix.accounts[3];
-                let _dest_user_pool_info = &ix.accounts[4];
-                let _manager_fee_info = &ix.accounts[5];
-                let _referrer_fee_info = &ix.accounts[6];
-                let pool_mint_info = &ix.accounts[7];
-
-                if let Some(mut lsts) = instruction.lsts.clone() {
-                    if let Some(alert_config) = lsts.get_mut(&pool_mint_info.pubkey.to_string()) {
-                        self.sort_thresholds(alert_config.thresholds.as_mut());
-                        for threshold in alert_config.thresholds.iter() {
-                            if *amount >= threshold.value {
-                                self.dispatch_platform_notifications(
-                                    &threshold.notification.destinations,
-                                    &threshold.notification.description,
-                                    Some(*amount),
-                                    Some("SOL"),
-                                    &parser.transaction_signature,
-                                )
-                                .await?;
-                                break;
-                            }
-                        }
-                    }
-                }
-            }
-            SplStakePoolProgram::WithdrawSol { ix, amount } => {
-                let _stake_pool_info = &ix.accounts[0];
-                let _withdraw_authority_info = &ix.accounts[1];
-                let _user_transfer_authority_info = &ix.accounts[2];
-                let _burn_from_pool_info = &ix.accounts[3];
-                let _reserve_stake_info = &ix.accounts[4];
-                let _destination_lamports_info = &ix.accounts[5];
-                let _manager_fee_info = &ix.accounts[6];
-                let pool_mint_info = &ix.accounts[7];
-
-                if let Some(mut lsts) = instruction.lsts.clone() {
-                    if let Some(alert_config) = lsts.get_mut(&pool_mint_info.pubkey.to_string()) {
-                        self.sort_thresholds(alert_config.thresholds.as_mut());
-                        for threshold in alert_config.thresholds.iter() {
-                            if *amount >= threshold.value {
-                                self.dispatch_platform_notifications(
-                                    &threshold.notification.destinations,
-                                    &threshold.notification.description,
-                                    Some(*amount),
-                                    Some("SOL"),
-                                    &parser.transaction_signature,
-                                )
-                                .await?;
-                                break;
-                            }
-                        }
-                    }
-                }
-            }
-            SplStakePoolProgram::DecreaseValidatorStakeWithReserve { ix, amount } => {
-                let stake_pool_info = &ix.accounts[0];
-                let _staker_info = &ix.accounts[1];
-                let _stake_pool_withdraw_authority_info = &ix.accounts[2];
-                let _validator_list_info = &ix.accounts[3];
-                let _reserve_stake_account_info = &ix.accounts[4];
-                let _validator_stake_info = &ix.accounts[5];
-                let _transient_stake_info = &ix.accounts[6];
-                let _clock_info = &ix.accounts[7];
-                let _stake_history_info = &ix.accounts[8];
-                let _system_program_info = &ix.accounts[9];
-                let _stake_program_info = &ix.accounts[10];
-
-                if let Some(mut stake_pools) = instruction.stake_pools.clone() {
-                    if let Some(alert_config) =
-                        stake_pools.get_mut(&stake_pool_info.pubkey.to_string())
-                    {
-                        self.sort_thresholds(alert_config.thresholds.as_mut());
-                        for threshold in alert_config.thresholds.iter() {
-                            if *amount > threshold.value {
-                                self.dispatch_platform_notifications(
-                                    &threshold.notification.destinations,
-                                    &threshold.notification.description,
-                                    Some(*amount),
-                                    Some("SOL"),
-                                    &parser.transaction_signature,
-                                )
-                                .await?;
-                                break;
-                            }
-                        }
-                    }
-                }
-            }
-            SplStakePoolProgram::Initialize
-            | SplStakePoolProgram::AddValidatorToPool
-            | SplStakePoolProgram::RemoveValidatorFromPool
-            | SplStakePoolProgram::DecreaseValidatorStake
-            | SplStakePoolProgram::SetPreferredValidator
-            | SplStakePoolProgram::UpdateValidatorListBalance
-            | SplStakePoolProgram::UpdateStakePoolBalance
-            | SplStakePoolProgram::CleanupRemovedValidatorEntries
-            | SplStakePoolProgram::SetManager
-            | SplStakePoolProgram::SetFee
-            | SplStakePoolProgram::SetStaker
-            | SplStakePoolProgram::SetFundingAuthority
-            | SplStakePoolProgram::CreateTokenMetadata
-            | SplStakePoolProgram::UpdateTokenMetadata
-            | SplStakePoolProgram::IncreaseAdditionalValidatorStake
-            | SplStakePoolProgram::DecreaseAdditionalValidatorStake
-            | SplStakePoolProgram::Redelegate
-            | SplStakePoolProgram::DepositStakeWithSlippage
-            | SplStakePoolProgram::WithdrawStakeWithSlippage
-            | SplStakePoolProgram::DepositSolWithSlippage
-            | SplStakePoolProgram::WithdrawSolWithSlippage => {
-                unreachable!()
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Handle Jito Vault Program
-    ///
-    /// - Notify only once for the first matching threshold.
-    async fn handle_jito_vault_program(
-        &mut self,
-        parser: &JitoTransactionParser,
-        jito_vault_program: &JitoVaultProgram,
-        instruction: &Instruction,
-    ) -> Result<(), JitoBellError> {
-        debug!("Jito Vault Program: {}", jito_vault_program);
-
-        match jito_vault_program {
-            JitoVaultProgram::MintTo { ix, min_amount_out } => {
-                let _config_info = &ix.accounts[0];
-                let _vault_info = &ix.accounts[1];
-                let vrt_mint_info = &ix.accounts[2];
-                let _depositor_info = &ix.accounts[3];
-                let _depositor_token_account = &ix.accounts[4];
-                let _vault_token_account = &ix.accounts[5];
-                let _depositor_vrt_token_account = &ix.accounts[6];
-                let _vault_fee_token_account = &ix.accounts[7];
-
-                if let Some(vrts) = instruction.vrts.clone() {
-                    if let Some((address, vrt_config)) =
-                        vrts.get_key_value(&vrt_mint_info.pubkey.to_string())
-                    {
-                        let vrt = Pubkey::from_str(address).unwrap();
-                        let divisor = self.divisor(&vrt).await;
-                        let symbol = self.vrt_symbol(&vrt).await;
-
-                        let mut thresholds = vrt_config.thresholds.clone();
-                        self.sort_thresholds(&mut thresholds);
-                        for threshold in vrt_config.thresholds.iter() {
-                            let min_amount_out = *min_amount_out as f64 / divisor;
-                            if min_amount_out >= threshold.value {
-                                self.dispatch_platform_notifications(
-                                    &threshold.notification.destinations,
-                                    &threshold.notification.description,
-                                    Some(min_amount_out),
-                                    Some(&symbol),
-                                    &parser.transaction_signature,
-                                )
-                                .await?;
-                                break;
-                            }
-                        }
-                    }
-                }
-            }
-            JitoVaultProgram::EnqueueWithdrawal { ix, amount } => {
-                let _config_info = &ix.accounts[0];
-                let vault_info = &ix.accounts[1];
-                let _vault_staker_withdrawal_ticket_info = &ix.accounts[2];
-                let _vault_staker_withdrawal_ticket_token_account_info = &ix.accounts[3];
-                let _staker_info = &ix.accounts[4];
-                let _staker_vrt_token_account_info = &ix.accounts[5];
-                let _base_info = &ix.accounts[6];
-
-                let vault_acc = self.rpc_client.get_account(&vault_info.pubkey).await?;
-                let vault = Vault::deserialize(&mut vault_acc.data.as_slice())?;
-
-                // VRT amount
-                if let Some(ref vrts) = instruction.vrts {
-                    if let Some((address, vrt_config)) =
-                        vrts.get_key_value(&vault.vrt_mint.to_string())
-                    {
-                        let vrt = Pubkey::from_str(address).unwrap();
-                        let divisor = self.divisor(&vrt).await;
-                        let symbol = self.vrt_symbol(&vrt).await;
-
-                        let mut thresholds = vrt_config.thresholds.clone();
-                        self.sort_thresholds(&mut thresholds);
-                        for threshold in vrt_config.thresholds.iter() {
-                            let amount = *amount as f64 / divisor;
-                            if amount >= threshold.value {
-                                self.dispatch_platform_notifications(
-                                    &threshold.notification.destinations,
-                                    &threshold.notification.description,
-                                    Some(amount),
-                                    Some(&symbol),
-                                    &parser.transaction_signature,
-                                )
-                                .await?;
-                                break;
-                            }
-                        }
-
-                        // USD amount
-                        if !vrt_config.usd_thresholds.is_empty() {
-                            let client = DefiLlamaClient::new();
-                            let vrt = Token::new(Chain::Solana, vrt.to_string());
-                            let prices = client.get_price(&vrt).await?;
-
-                            if let Some(usd_price) = prices.coins.values().last() {
-                                let mut sorted_usd_thresholds = vrt_config.usd_thresholds.clone();
-                                sorted_usd_thresholds.sort_by(|a, b| {
-                                    b.value
-                                        .partial_cmp(&a.value)
-                                        .unwrap_or(std::cmp::Ordering::Equal)
-                                });
-
-                                for usd_threshold in sorted_usd_thresholds.iter() {
-                                    let amount = *amount as f64 / 1_000_000_000_f64;
-                                    let amount = (amount * usd_price.price) as u64;
-
-                                    if amount >= usd_threshold.value {
-                                        self.dispatch_platform_notifications(
-                                            &usd_threshold.notification.destinations,
-                                            &usd_threshold.notification.description,
-                                            Some(amount as f64),
-                                            Some("USD"),
-                                            &parser.transaction_signature,
-                                        )
-                                        .await?;
-                                        break;
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-            JitoVaultProgram::InitializeConfig
-            | JitoVaultProgram::InitializeVault
-            | JitoVaultProgram::InitializeVaultWithMint
-            | JitoVaultProgram::InitializeVaultOperatorDelegation
-            | JitoVaultProgram::InitializeVaultNcnTicket
-            | JitoVaultProgram::InitializeVaultNcnSlasherOperatorTicket
-            | JitoVaultProgram::InitializeVaultNcnSlasherTicket
-            | JitoVaultProgram::WarmupVaultNcnTicket
-            | JitoVaultProgram::CooldownVaultNcnTicket
-            | JitoVaultProgram::WarmupVaultNcnSlasherTicket
-            | JitoVaultProgram::CooldownVaultNcnSlasherTicket
-            | JitoVaultProgram::ChangeWithdrawalTicketOwner
-            | JitoVaultProgram::BurnWithdrawalTicket
-            | JitoVaultProgram::SetDepositCapacity
-            | JitoVaultProgram::SetFees
-            | JitoVaultProgram::SetProgramFee
-            | JitoVaultProgram::SetProgramFeeWallet
-            | JitoVaultProgram::SetIsPaused
-            | JitoVaultProgram::DelegateTokenAccount
-            | JitoVaultProgram::SetAdmin
-            | JitoVaultProgram::SetSecondaryAdmin
-            | JitoVaultProgram::AddDelegation
-            | JitoVaultProgram::CooldownDelegation
-            | JitoVaultProgram::UpdateVaultBalance
-            | JitoVaultProgram::InitializeVaultUpdateStateTracker
-            | JitoVaultProgram::CrankVaultUpdateStateTracker
-            | JitoVaultProgram::CloseVaultUpdateStateTracker
-            | JitoVaultProgram::CreateTokenMetadata
-            | JitoVaultProgram::UpdateTokenMetadata
-            | JitoVaultProgram::SetConfigAdmin => {
-                unreachable!()
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Sends a notification for each matching `CopyDirectedStakeTargets` instruction
-    /// that includes notification metadata.
-    async fn handle_jito_steward_program(
-        &mut self,
-        parser: &JitoTransactionParser,
-        jito_steward_instruction: &JitoStewardInstruction,
-        instruction: &Instruction,
-    ) -> Result<(), JitoBellError> {
-        debug!("Jito Steward Instruction: {jito_steward_instruction}");
-
-        if let JitoStewardInstruction::CopyDirectedStakeTargets {
-            ix: _,
-            vote_pubkey: _,
-            total_target_lamports,
-            validator_list_index: _,
-        } = jito_steward_instruction
-        {
-            if let Some(ref notification_info) = instruction.notification_info {
-                self.dispatch_platform_notifications(
-                    &notification_info.destinations,
-                    &notification_info.description,
-                    Some(*total_target_lamports as f64),
-                    Some("lamports"),
-                    &parser.transaction_signature,
-                )
-                .await?;
-            }
-        }
-
-        Ok(())
-    }
-
     /// Dispatch platform notifications
     ///
     /// - Return error only if ALL platforms failed, or handle as needed
-    async fn dispatch_platform_notifications(
+    pub(crate) async fn dispatch_platform_notifications(
         &mut self,
         destinations: &[Destination],
         description: &str,
